examples/dreambooth/requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
peft==0.7.0
# diffusers==0.27.0.dev0 # using the source code to install: > pip install -e .